Update newnew-api dependency, implement hiding own Posts

+++ components/organisms/decision/PostModal.tsx
@@ -181,6 +181,11 @@ const PostModal: React.FunctionComponent<IPostModal> = ({
     [postStatus]
   );
 
+  const postStatusesToUseHideInsteadOfDelete = useMemo(
+    () => ['succeeded', 'failed', 'deleted_by_creator', 'deleted_by_admin'],
+    []
+  );
+
   const shouldRenderVotingFinishedModal = useMemo(
     () =>
       postStatus === 'succeeded' ||
@@ -293,22 +298,43 @@ const PostModal: React.FunctionComponent<IPostModal> = ({
 
   const handleDeletePost = useCallback(async () => {
     try {
-      const payload = new newnewapi.DeleteMyPostRequest({
-        postUuid: postParsed?.postUuid,
-      });
+      // "Hide" Post instead of deleting it
+      if (postStatusesToUseHideInsteadOfDelete.includes(postStatus)) {
+        const payload = new newnewapi.MarkPostRequest({
+          postUuid: postParsed?.postUuid,
+          markAs: newnewapi.MarkPostRequest.Kind.HIDDEN,
+        });
 
-      const res = await deleteMyPost(payload);
+        const res = await markPost(payload);
 
-      if (!res.error) {
-        console.log('Post deleted/cancelled');
-        handleUpdatePostStatus('DELETED_BY_CREATOR');
-        handleRemovePostFromState?.();
-        handleCloseDeletePostModal();
+        if (!res.error) {
+          handleUpdatePostStatus('DELETED_BY_CREATOR');
+          handleRemovePostFromState?.();
+          handleCloseDeletePostModal();
+        }
+      } else {
+        const payload = new newnewapi.DeleteMyPostRequest({
+          postUuid: postParsed?.postUuid,
+        });
+
+        const res = await deleteMyPost(payload);
+
+        if (!res.error) {
+          handleUpdatePostStatus('DELETED_BY_CREATOR');
+          handleRemovePostFromState?.();
+          handleCloseDeletePostModal();
+        }
       }
     } catch (err) {
       console.error(err);
     }
-  }, [handleRemovePostFromState, handleUpdatePostStatus, postParsed?.postUuid]);
+  }, [
+    handleRemovePostFromState,
+    handleUpdatePostStatus,
+    postParsed?.postUuid,
+    postStatus,
+    postStatusesToUseHideInsteadOfDelete,
+  ]);
 
   useEffect(() => {
     if (commentIdFromUrl) {
@@ -904,7 +930,7 @@ const PostModal: React.FunctionComponent<IPostModal> = ({
           <PostEllipseMenuModeration
             postType={typeOfPost as string}
             isVisible={ellipseMenuOpen}
-            canDeletePost={postStatus !== 'failed'}
+            canDeletePost
             handleClose={handleEllipseMenuClose}
             handleOpenDeletePostModal={handleOpenDeletePostModal}
             anchorElement={moreButtonRef.current}
@@ -914,7 +940,7 @@ const PostModal: React.FunctionComponent<IPostModal> = ({
           <PostEllipseModalModeration
             postType={typeOfPost as string}
             zIndex={11}
-            canDeletePost={postStatus !== 'failed'}
+            canDeletePost
             isOpen={ellipseMenuOpen}
             onClose={handleEllipseMenuClose}
             handleOpenDeletePostModal={handleOpenDeletePostModal}
@@ -939,7 +965,6 @@ const PostModal: React.FunctionComponent<IPostModal> = ({
     handleShareClose,
     isMobile,
     postParsed?.postUuid,
-    postStatus,
     shareMenuOpen,
     theme.colors.dark,
     theme.colors.white,
